+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Version 0.2.0
+
+- Renamed from JSONConvertible -> Convertible
+- Minor improvements and refactoring
+- Added example project
+
+## Version 0.1.1
+
+- Renamed from JSON -> JSONConvertible
+- Minor improvements
+
 ## Version 0.1.0
 
 - Initial version